backend/requirements.txt
flask
requests
beautifulsoup4
selenium
pdfminer.six
python-docx
tldextract
transformers